+++ lib/screens/track_metadata_screen.dart
@@ -938,11 +938,9 @@ class _TrackMetadataScreenState extends ConsumerState<TrackMetadataScreen> {
       return;
     }
     
-    await SharePlus.instance.share(
-      ShareParams(
-        files: [XFile(item.filePath)],
-        text: '${item.trackName} - ${item.artistName}',
-      ),
+    await Share.shareXFiles(
+      [XFile(item.filePath)],
+      text: '${item.trackName} - ${item.artistName}',
     );
   }